requirements/local.txt
# Local development dependencies go here
-r base.txt

django-extensions==1.7.4
Werkzeug==0.11.11
django-debug-toolbar==1.5